requirements.txt
PyQt6
Jinja2
python-docx
weasyprint
pytest
